Type fixes.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@384 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/wordlists.c
@@ -214,7 +214,7 @@ bool configure_wordlist(const char *val)
     *tmp++ = '\0';
     while (isspace(*tmp)) tmp += 1;
 
-    override=atof(tmp);
+    override=atoi(tmp);
     tmp += strcspn(tmp, ", \t");
     *tmp++ = '\0';
     while (isspace(*tmp)) tmp += 1;
